♻️ Rename test project from ProductController.Tests to ProductList.Tests and update solution references
