+++ ProductList.Tests/ProductList.Tests.csproj
@@ -20,8 +20,8 @@
     <Using Include="Xunit" />
   </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\ProductList.csproj" />
-  </ItemGroup>
+<ItemGroup>
+    <ProjectReference Include="..\ProductList\ProductList.csproj" />
+</ItemGroup>
 
 </Project>
